integrate DrawerMenu into main app
