+++ src/frontend/routes/MainRoutes.tsx
@@ -1,7 +1,7 @@
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import React from 'react';
 import { Screens } from '../helpers';
-import { Chat, Drawer } from '../screens';
+import { Chat, DrawerMenu } from '../screens';
 
 export type MainDrawerParams = {
   [Screens.Chat]: undefined;
@@ -11,7 +11,7 @@ const MainRouteDrawer = createDrawerNavigator<MainDrawerParams>();
 
 export function MainRoutes() {
   return (
-    <MainRouteDrawer.Navigator drawerContent={(props) => <Drawer />}>
+    <MainRouteDrawer.Navigator drawerContent={(props) => <DrawerMenu />}>
       <MainRouteDrawer.Screen key={'123'} name={Screens.Chat} component={Chat} />
     </MainRouteDrawer.Navigator>
   );
