--- src/frontend/screens/Drawer/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-
-export function Drawer() {
-  return (
-    <View>
-      <Text>Drawer</Text>
-    </View>
-  );
-}
