+++ src/frontend/screens/index.ts
@@ -2,7 +2,7 @@ export * from './Fallback';
 export * from './Landing';
 export * from './Login';
 export * from './Chat';
-export * from './Drawer';
+export * from './DrawerMenu';
 export * from './SignUp';
 export * from './ForgotPassword';
 export * from './ResetPassword';
